+++ packages/blinkcard-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microblink/blinkcard-core
 
+## 3000.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/blinkcard-wasm@3000.0.8
+  - @microblink/blinkcard-worker@3000.0.8
+
 ## 3000.0.7
 
 ### Patch Changes