+++ packages/blinkcard-worker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @microblink/blinkcard-worker
 
+## 3000.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/blinkcard-wasm@3000.0.8
+
 ## 3000.0.7
 
 ### Patch Changes